// GESTIÓN DE CALIDAD SOFTWARE/Taller Ionic/appIonicEntrega/src/app/models/starship.ts
export class Starship {
    name: string = "";
    model: string = "";
    manufacturer: string = "";
    cost_in_credits: number = 0;
    length: number = 0;
    crew: number = 0;
    passengers: number = 0;
    cargo_capacity: number = 0;
    hyperdrive_rating: number = 0;
    starship_class: string = "";
  }
  